Fix error page style
